[#1742] Remove redundant iceoryx2-services-common crate
