+++ BUILD.bazel
@@ -329,7 +329,6 @@ filegroup(
         "//iceoryx2-pal/posix:all_srcs",
         "//iceoryx2-pal/print:all_srcs",
         "//iceoryx2-pal/testing:all_srcs",
-        "//iceoryx2-services/common:all_srcs",
         "//iceoryx2-services/discovery:all_srcs",
         "//iceoryx2-gateway/gateway:all_srcs",
         "//iceoryx2-gateway/testing:all_srcs",
