+++ examples/Cargo.toml
@@ -12,7 +12,6 @@ version = { workspace = true }
 
 [dependencies]
 iceoryx2-log = { workspace = true, features = ["std"] }
-iceoryx2-services-common = { workspace = true, features = ["std"] }
 iceoryx2-services-discovery = { workspace = true, features = ["std"] }
 iceoryx2 = { workspace = true, features = ["std"] }
 iceoryx2-cal = { workspace = true, features = ["std"] }
