+++ iceoryx2-cli/Cargo.toml
@@ -41,7 +41,6 @@ path = "lib/src/lib.rs"
 [dependencies]
 iceoryx2-log = { workspace = true, features = ["std"] }
 iceoryx2-userland-record-and-replay = { workspace = true }
-iceoryx2-services-common = { workspace = true, features = ["std"] }
 iceoryx2-services-discovery = { workspace = true, features = ["std"] }
 iceoryx2 = { workspace = true, features = ["std"] }
 iceoryx2-cal = { workspace = true, features = ["std"] }
