+++ iceoryx2-gateway/backend/BUILD.bazel
@@ -29,7 +29,6 @@ rust_library(
         "//conditions:default": [],
     }),
     deps = [
-        "//iceoryx2-services/common:iceoryx2-services-common",
         "//iceoryx2:iceoryx2",
         "//iceoryx2-bb/posix:iceoryx2-bb-posix",
         "//iceoryx2-bb/testing:iceoryx2-bb-testing",
