+++ iceoryx2-gateway/conformance-tests/Cargo.toml
@@ -17,7 +17,6 @@ std = [
   "iceoryx2-bb-posix/std",
   "iceoryx2-bb-testing/std",
   "iceoryx2-bb-testing-macros/std",
-  "iceoryx2-services-common/std",
   "iceoryx2-services-discovery/std",
   "iceoryx2-gateway/std",
   "iceoryx2-gateway-backend/std",
@@ -33,7 +32,6 @@ iceoryx2 = { workspace = true }
 iceoryx2-bb-posix = { workspace = true }
 iceoryx2-bb-testing = { workspace = true }
 iceoryx2-bb-testing-macros = { workspace = true }
-iceoryx2-services-common = { workspace = true }
 iceoryx2-services-discovery = { workspace = true }
 iceoryx2-gateway = { workspace = true }
 iceoryx2-gateway-backend = { workspace = true }
