+++ iceoryx2-gateway/gateway/BUILD.bazel
@@ -29,7 +29,6 @@ rust_library(
         "//conditions:default": [],
     }),
     deps = [
-        "//iceoryx2-services/common:iceoryx2-services-common",
         "//iceoryx2-gateway/backend:iceoryx2-gateway-backend",
         "//iceoryx2-services/discovery:iceoryx2-services-discovery",
         "//iceoryx2:iceoryx2",
