+++ iceoryx2-gateway/gateway/Cargo.toml
@@ -15,7 +15,6 @@ version = { workspace = true }
 default = ["std"]
 std = [
   "iceoryx2-log/std",
-  "iceoryx2-services-common/std",
   "iceoryx2-services-discovery/std",
   "iceoryx2-gateway-backend/std",
   "iceoryx2/std",
@@ -26,6 +25,5 @@ std = [
 iceoryx2-log = { workspace = true }
 iceoryx2 = { workspace = true }
 iceoryx2-bb-container = { workspace = true }
-iceoryx2-services-common = { workspace = true }
 iceoryx2-services-discovery = { workspace = true }
 iceoryx2-gateway-backend = { workspace = true }
