+++ iceoryx2-gateway/testing/BUILD.bazel
@@ -34,7 +34,6 @@ rust_library(
         "//iceoryx2-bb/elementary:iceoryx2-bb-elementary",
         "//iceoryx2-bb/posix:iceoryx2-bb-posix",
         "//iceoryx2-bb/system-types:iceoryx2-bb-system-types",
-        "//iceoryx2-services/common:iceoryx2-services-common",
         "//iceoryx2-gateway/backend:iceoryx2-gateway-backend",
         "@crate_index//:postcard",
         "@crate_index//:serde",
