--- iceoryx2-services/common/BUILD.bazel
@@ -1,29 +1,0 @@
-# Copyright (c) 2025 Contributors to the Eclipse Foundation
-#
-# See the NOTICE file(s) distributed with this work for additional
-# information regarding copyright ownership.
-#
-# This program and the accompanying materials are made available under the
-# terms of the Apache Software License 2.0 which is available at
-# https://www.apache.org/licenses/LICENSE-2.0, or the MIT license
-# which is available at https://opensource.org/licenses/MIT.
-#
-# SPDX-License-Identifier: Apache-2.0 OR MIT
-
-package(default_visibility = ["//visibility:public"])
-
-load("@rules_rust//rust:defs.bzl", "rust_library")
-
-filegroup(
-    name = "all_srcs",
-    srcs = glob(["**"]),
-)
-
-rust_library(
-    name = "iceoryx2-services-common",
-    srcs = glob(["src/**/*.rs"]),
-    deps = [
-        "//iceoryx2",
-        "@crate_index//:serde",
-    ],
-)
