--- iceoryx2-services/common/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "iceoryx2-services-common"
-description = "iceoryx2-services: common types and utilities"
-categories = { workspace = true }
-edition = { workspace = true }
-homepage = { workspace = true }
-keywords = { workspace = true }
-license = { workspace = true }
-readme = "README.md"
-repository = { workspace = true }
-rust-version = { workspace = true }
-version = { workspace = true }
-
-[features]
-default = []
-std = [
-  "iceoryx2/std",
-]
-
-[dependencies]
-iceoryx2 = { workspace = true }
-
-serde = { workspace = true }
