--- iceoryx2-services/common/README.md
@@ -1,3 +1,0 @@
-# iceoryx2-services-common
-
-Common types and utilities shared across the `iceoryx2-services` crates.
