--- iceoryx2-services/common/src/lib.rs
@@ -1,65 +1,0 @@
-// Copyright (c) 2026 Contributors to the Eclipse Foundation
-//
-// See the NOTICE file(s) distributed with this work for additional
-// information regarding copyright ownership.
-//
-// This program and the accompanying materials are made available under the
-// terms of the Apache Software License 2.0 which is available at
-// https://www.apache.org/licenses/LICENSE-2.0, or the MIT license
-// which is available at https://opensource.org/licenses/MIT.
-//
-// SPDX-License-Identifier: Apache-2.0 OR MIT
-
-//! # iceoryx2-services-common
-//!
-//! Common types and utilities shared across the `iceoryx2-services` crates.
-
-#![no_std]
-
-use iceoryx2::{
-    prelude::ZeroCopySend,
-    service::{service_hash::ServiceHash, static_config::StaticConfig},
-};
-
-extern crate alloc;
-
-/// Communicates discovered changes to services in the system.
-///
-/// Owning variant. Can be used as shared memory payload.
-#[derive(Clone, Debug, ZeroCopySend, serde::Serialize, serde::Deserialize)]
-// Largest variant dictates the size required for ZeroCopySend payload.
-#[allow(clippy::large_enum_variant)]
-#[repr(C)]
-pub enum DiscoveryEvent {
-    /// A service has been added to the system.
-    ///
-    /// Contains the static configuration of the newly added service.
-    Added(StaticConfig),
-
-    /// A service has been removed from the system.
-    ///
-    /// Contains the hash identifying the removed service.
-    Removed(ServiceHash),
-}
-
-/// Communicates discovered changes to services in the system.
-///
-/// Non-owning variant. Used to propagate through API layers without
-/// copy.
-#[derive(Debug)]
-pub enum DiscoveryEventRef<'a> {
-    /// A service has been added to the system.
-    Added(&'a StaticConfig),
-
-    /// A service has been removed from the system.
-    Removed(&'a ServiceHash),
-}
-
-impl<'a> From<&'a DiscoveryEvent> for DiscoveryEventRef<'a> {
-    fn from(event: &'a DiscoveryEvent) -> Self {
-        match event {
-            DiscoveryEvent::Added(sc) => DiscoveryEventRef::Added(sc),
-            DiscoveryEvent::Removed(h) => DiscoveryEventRef::Removed(h),
-        }
-    }
-}
