+++ iceoryx2-services/discovery/BUILD.bazel
@@ -29,7 +29,6 @@ rust_library(
         "//conditions:default": [],
     }),
     deps = [
-        "//iceoryx2-services/common:iceoryx2-services-common",
         "//iceoryx2",
         "//iceoryx2-bb/concurrency:iceoryx2-bb-concurrency",
         "//iceoryx2-bb/container:iceoryx2-bb-container",
