+++ iceoryx2-services/discovery/Cargo.toml
@@ -22,11 +22,9 @@ std = [
   "iceoryx2-bb-container/std",
   "iceoryx2-bb-elementary/std",
   "iceoryx2/std",
-  "iceoryx2-services-common/std",
 ]
 
 [dependencies]
-iceoryx2-services-common = { workspace = true }
 iceoryx2 = { workspace = true }
 iceoryx2-bb-concurrency = { workspace = true }
 iceoryx2-bb-container = { workspace = true }
