+++ integrations/ros2/gateway-backend/Cargo.toml
@@ -14,7 +14,6 @@ version = { workspace = true }
 iceoryx2 = { workspace = true }
 iceoryx2-bb-concurrency = { workspace = true }
 iceoryx2-log = { workspace = true }
-iceoryx2-services-common = { workspace = true }
 iceoryx2-gateway-backend = { workspace = true }
 
 libloading = { workspace = true }
