+++ integrations/zenoh/gateway-backend/Cargo.toml
@@ -18,7 +18,6 @@ harness = false
 iceoryx2 = { workspace = true, features = ["std"] }
 iceoryx2-log = { workspace = true, features = ["std"] }
 iceoryx2-bb-concurrency = { workspace = true, features = ["std"] }
-iceoryx2-services-common = { workspace = true, features = ["std"] }
 iceoryx2-gateway-backend = { workspace = true, features = ["std"] }
 
 flume = { workspace = true }
